third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/gc-after-body-removed-expected.txt

CONSOLE ERROR: line 5: Uncaught TypeError: detail.createShadowRoot is not a function
